# internal/apidocs/openapi.yaml
openapi: 3.0.3
info:
  title: Donna API
  description: |
    HTTP and WebSocket API for the Donna voice backend (`donna-server-go`).

    ## Authentication

    Protected routes require a Supabase access token (JWT) in the `Authorization` header:

    ```
    Authorization: Bearer <access_token>
    ```

    The `/voice` WebSocket accepts the same token as a query parameter when auth is enabled:

    ```
    ws://host:8787/voice?token=<access_token>
    ```

    When `SUPABASE_URL` is unset (local dev), auth is disabled and tokens are optional.

    ## Voice WebSocket protocol

    After connecting to `/voice`, exchange JSON text frames using the schemas in
    **VoiceClientMessage** and **VoiceServerMessage** below. Voice is
    speech-to-text only; talk-mode clients send the transcript through
    `POST /chat` for the shared text chat harness. Typical talk-mode flow:

    1. Client sends `session.start` (`mode: talk`)
    2. Server responds with `session.ready` and `turn.phase: idle`
    3. Client streams `audio.chunk` frames (base64 PCM16)
    4. Client sends `turn.end` to trigger STT
    5. Server emits `turn.transcript`, then `turn.done`
    6. Client sends `session.end`, then posts the transcript to `/chat`
    7. Notes mode (`mode: notes`) still creates a note from the transcript on `turn.done`
  version: 1.0.0
  contact:
    name: Donna
servers:
  - url: http://localhost:8787
    description: Local development
  - url: https://donna-server-production.up.railway.app
...
paths:
...
  /knowledge/ingest:
    post:
      tags: [Knowledge]
      summary: Ingest knowledge
      description: |
        Ingest a URL, plain-text note, or file into the user's knowledge base.
        Content is extracted, stored, and queued for compilation into facts.

        Accepts either `application/json` (url or text) or `multipart/form-data` (file upload, max 15 MB).
      operationId: postKnowledgeIngest
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestJsonRequest"
            examples:
              url:
                summary: Ingest a URL
                value:
                  url: https://example.com/article
              text:
                summary: Ingest a text note
                value:
                  text: Meeting notes from today...
                  title: Team sync
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Document, image, or audio file (see GET /knowledge/formats)
      responses:
        "200":
          description: Ingest queued successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestSuccessResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: missing_token
        "413":
          description: File too large (> 15 MB)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: file_too_large
        "415":
          description: Unsupported media type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: unsupported_content_type
        "422":
          description: Invalid request body or extraction failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Knowledge base persistence disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: knowledge_disabled
...
  /account:
    get:
      tags: [Account]
      summary: Get account preferences
      description: Returns the authenticated user's LLM model preference and available models.
      operationId: getAccount
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Account preferences
          content:
            application/json:
              schema:
                type: object
                properties:
                  llm_model:
                    type: string
                  available_models:
                    type: array
                    items:
                      type: string
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags: [Account]
      summary: Update account preferences
      operationId: updateAccount
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [llm_model]
              properties:
                llm_model:
                  type: string
      responses:
        "200":
          description: Preferences updated
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Account]
      summary: Delete account
      description: |
        Permanently deletes the authenticated user's data: storage objects,
        knowledge base rows, conversations, and the Supabase auth user.
      operationId: deleteAccount
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteAccountResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: missing_token
        "500":
          description: Deletion failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: delete_failed
                message: account deletion unavailable
  /account/export:
    get:
      tags: [Account]
      summary: Download account data export
      description: |
        Returns a ZIP archive of the authenticated user's data: voice and text
        conversations (transcripts and audio), notes, memory facts, ingested
        knowledge sources and files, profile, preferences, and compile log.

        Rate limited to one export per 15 minutes per user.
      operationId: exportAccount
      security:
        - bearerAuth: []
      responses:
        "200":
          description: ZIP data export
          content:
            application/zip:
              schema:
                type: string
                format: binary
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Export rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Export unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /voice:
    get:
      tags: [Voice]
      summary: Voice WebSocket
      description: |
        Upgrades to a WebSocket connection for real-time voice interaction (STT → LLM → TTS).

        **Connection:** `ws://host:8787/voice?token=<jwt>` (token required when auth is enabled)

        **Subprotocol:** none — messages are JSON text frames.

        See component schemas **VoiceClientMessage** and **VoiceServerMessage** for the wire format.
      operationId: connectVoice
      parameters:
        - name: token
          in: query
          required: false
          description: Supabase JWT (required when auth is enabled)
          schema:
            type: string
      responses:
        "101":
          description: WebSocket upgrade successful
        "401":
          description: WebSocket closed with code 4401 (missing or invalid token)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase access token
  schemas:
...
    IngestJsonRequest:
      type: object
      description: Provide either `url` or `text` (not both).
      properties:
        url:
          type: string
          format: uri
          description: Web page or document URL to fetch and extract
        text:
          type: string
          description: Plain-text note content
        title:
          type: string
          description: Optional title for text notes
    IngestSuccessResponse:
      type: object
      required: [source_id, asset_kind, title, status]
      properties:
        source_id:
          type: string
          format: uuid
        asset_kind:
          type: string
          example: text
        title:
          type: string
          nullable: true
        status:
          type: string
          enum: [queued]
          example: queued
    DeleteAccountResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
...